Make mount order deterministic
Protobuf makes no guarantees about the order of map entries and may even randomize in certain build configurations. This can lead to unexpected behavior in sandbox2, as the order of mounts can affect the behavior of the sandboxed application.

This CL adds an `index` field to the `MountTree` proto. This field is used to keep track of the insertion order of mounts. When mounts are created, they are sorted by their index. This ensures that the order of mounts is always deterministic, regardless of the build configuration.

PiperOrigin-RevId: 691329053
Change-Id: Ib73504b0ced8a00e15a68c3e85af5d542fdd8439