Add a BPF evaluator for sandbox2
This will be needed for unotify monitor to do dynamic policy decisions.
All DENY instructions in the original filter are replaced with unotify and so are all TRACE instructions. Unfortunately we do not have even a single bit of info we can pass from the filter to unotify, so we'll need to evaluate the original filter to differentiate between those 2 cases.

PiperOrigin-RevId: 691411537
Change-Id: I87aeb4af4a2ff143af181093b8f7249fe2bcaa4e