Bazel: Use `llvm-config` to find system LLVM
This is a much more reliable way to find system LLVM than the previous method, which was based on guessing the location of the LLVM include directory.

The new method also allows us to get the list of system libraries that need to be linked against, which is necessary for some targets.

On Debian, this supports all LLVM/Clang version from 11 up to (and including) 20 (once released).

Note: We don't need to do anything for CMake, as LLVM ships with CMake
config files that perform the same function.
PiperOrigin-RevId: 692190033
Change-Id: I4ffc1ca33739358b8c2b87335536a563ae134c72