clang_generator: Fix handling of named enum types
Previously, the function `GetQualTypeName` inconsistently handled the "enum"
keyword.  Specifically, it discarded the keyword when `clang::QualType`
was used, and retained it when `clang::TypeDecl` was used as a base.

This inconsistency led to discarding valid enums in `GetTypeDeclarations`
function.

This commit changes the method for generating names for enums to ensure
consistent treatment of the "enum" keyword.

PiperOrigin-RevId: 694829354
Change-Id: I96f3f5e9d812cc0d6c45a5c2c3b622799a54a178