Sandbox2: `UnotifyMonitor` fix possible UB with lambda deleter
The previous lambda deleter has a different type per translation unit it is
included in. And since `UnotifyMonitor` is not defined in an anonymous
namespace we may encounter undefined behavior. This fixes the
`subobject-linkage` compiler warning.

The reason we need a custom deleter in the first place is that the `res_` and
`resp_` members need to be allocated using `malloc()`.

If we need this elsewhere, we might move `StdFreeDeleter` (and others like it)
to a `util/deleters.h` utility.

PiperOrigin-RevId: 699102219
Change-Id: I4466a07cfa8a07d449a45690ba1c65eb6253d111