policybuilder: Add `AnchorPathAbsolute` utility function.
This utility function can be used by users of the `PolicyBuilder` to create an absolute path anchored to a defined `base`.

* If the path passed as `relative_path` is absolute; The path will be returned as is.
* If the passed path is relative and `base` is not provided; The path will be resolved relative to the current working directory.
* If the passed path is relative and an absolute `base` is provided; The path will
be resolved relative to `base`.
* If both, `relative_path` and `base` are relative, then first `base` will be
resolved relative to the current working directory, and then `relative_path` will be resolved relative to `base`.

In all cases where `relative_path` is relative; Any path elements such as dot-dot, dot, slash-slash, and trailing slash are cleaned up as string manipulation (no syscall is performed for path normalization). The result is then anchored to the base directory.

In case of an error, an empty path will be returned. Calling `AddFile`, `AddFileAt`, `AddDirectory`, and `AddDirectoryAt` with empty strings does not result in an error on policy compilation.

PiperOrigin-RevId: 719239498
Change-Id: Ie51cfb9171ee74934d2b663414c83b6245707aed