sandbox: Reduce log-spam for ungraceful SAPI sandbox termination.
In cases where `sapi::Sandbox::Terminate()` is called with `attempt_graceful_exit` set to false, a warning log will be written. This is due to the sandbox getting killed before the `AwaitResult()` call returns.

In this change, we reduce the log-spam for situations where the boolean is false and the result is `sandbox2::Result::EXTERNAL_KILL`.

PiperOrigin-RevId: 720516033
Change-Id: I4f59ac98bcd89b78c48c3cf21c8a2369726ae2ce