PtraceMonitor: Add a flag to use deadline manager instead of sigtimedwait
As SIGCHLD is process directed in case where there are many sandboxees running
it is not reliably delivered (any monitor can get it). Thus the whole relies
on the monitors actually polling the status every 500ms. This causes bigger
latency and unneccessary CPU load.

This change puts the feature behind a flag for easier rollout/rollbacks.

PiperOrigin-RevId: 721312381
Change-Id: I7151e9f308fd76e0103539dea93610c164e9db47