Pass more precise defines to the clang tool
The immediate problem at hand is that TSan/MSan change libc++ ABI
so that std namespace becomes std::__xsan, which in turn changes name mangling.
As the result all names we extract don't match the actual names in the final build.

Use the actual defines that would be used during C++ compilation for the clang tool.
This fixes the name mangling problem.

PiperOrigin-RevId: 792115834
Change-Id: Ia064dceb1d19d2b9291ec7bb1c1edf0807b9385e