Fix GetRemote not returning what SetRemote set in RemotePtr
`RemotePtr` overrides `Var`'s `SetRemote` but not `GetRemote`. That way, `RemotePtr` exposes an inconsistent API, because `GetRemote` does not retrieve what `SetRemote` set.

This CL adds a test for the above issue and fixes it by overriding `GetRemote` in the same way as `SetRemote` is overridden: delegating to `pointed_obj_`.

PiperOrigin-RevId: 821590955
Change-Id: I3f4396c9333d72f9eaa1c9abd4cbd778c62aae45